requirements.txt
scikit-learn==1.7.2
pandas==2.3.3
numpy==2.2.6
mlflow==3.10.1
pyyaml==6.0.3
scipy==1.15.3
requests==2.32.5
matplotlib==3.10.8
pyarrow==23.0.1